# openapi-ci-api.yaml
openapi: 3.0.3
info:
  title: CCExtractor CI System API
  version: 1.2.0
  description: |
    Security-hardened JSON-only REST API for the CCExtractor CI/sample platform.
    Designed for AI agents and CI automation. Enforces scoped Bearer token auth,
    strict input validation, rate limiting on all routes, and safe defaults
    throughout. No browser sessions, no HTML, no implicit permissions.

    **Authentication:** All endpoints require bearer token authentication unless
    explicitly marked with `security: []` (only /system/health and POST /auth/tokens).

    **Rate-limit headers:** Every response includes `X-RateLimit-Limit`,
    `X-RateLimit-Remaining`, and `X-RateLimit-Reset` headers. These are modeled
    explicitly on the 429 response for brevity; they are present on all responses
    regardless of status code.

  contact:
    name: CCExtractor Development
    url: https://github.com/CCExtractor/sample-platform
  license:
    name: GPL-3.0-only
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://sampleplatform.ccextractor.org/api/v1
    description: Production

#
# Global security: all endpoints require auth
# unless explicitly overridden with security: []
#
security:
  - bearerAuth: []
...
paths:
...
  /system/queue:
    get:
      tags: [System]
      summary: Get queue depth and currently running jobs
      operationId: getQueue
      security:
        - bearerAuth: []
      x-required-scope: system:read
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: platform
          in: query
          schema:
            type: string
            enum: [linux, windows]
        - name: status
          in: query
          schema:
            type: string
            enum: [queued, running]
      responses:
        "200":
          description: Queue status and active jobs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      queue_depth:
                        type: integer
                        minimum: 0
                      running_count:
                        type: integer
                        minimum: 0
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/QueueJob"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"
  /runs/{run_id}/artifacts:
    get:
      tags: [System]
      summary: List downloadable artifacts for a run
      operationId: listArtifacts
      description: >
        Only returns artifacts with a verified download_url from at least one
        storage backend. storage_status=degraded means one backend only;
        storage_status=missing means neither backend has the file (download_url
        will be null). Never returns a URL that has not been verified to exist.
      security:
        - bearerAuth: []
      x-required-scope: results:read
      parameters:
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: type
          in: query
          schema:
            type: string
            enum: [build_log, sample_output, expected_output, diff, media_info, binary, coredump, combined_stdout]
      responses:
        "200":
          description: Paginated run artifacts
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        default:
          $ref: "#/components/responses/Error"

#
# COMPONENTS
#
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: >
        Opaque server-side API token. Obtain via POST /auth/tokens.
        The CI worker token used by /ci/progress-reporter is a separate
        secret and is NOT valid here. Never use browser session cookies
        for API clients.

  # HEADERS
  headers:
    RateLimitLimit:
      description: Maximum requests allowed in the current window
      schema:
        type: integer
        example: 120
    RateLimitRemaining:
      description: Requests remaining in the current window
      schema:
        type: integer
        example: 117
    RateLimitReset:
      description: Unix timestamp when the rate limit window resets
      schema:
        type: integer
        example: 1748908800

  # PARAMETERS
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return (1–100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Offset:
      name: offset
      in: query
      description: Number of results to skip for pagination
      schema:
        type: integer
        minimum: 0
        maximum: 2147483647
        default: 0
...
    RunId:
      name: run_id
      in: path
      required: true
      description: Numeric run ID
      schema:
        type: integer
        minimum: 1
...
  responses:
    BadRequest:
      description: Request body or query parameters failed schema validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Request failed schema validation.
            details:
              fields:
                commit_sha: Must match pattern ^[a-fA-F0-9]{40}$
                platform: Must be one of [linux, windows]
    Unauthorized:
      description: Missing, expired, or invalid bearer token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Bearer token is missing, expired, or invalid.
            details: {}
    Forbidden:
      description: Token is valid but lacks the required scope or role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden
            message: Token does not have the required scope for this operation.
            details:
              required_scope: runs:write
              token_scopes: [runs:read, results:read]
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: Run 9317 not found.
            details:
              resource: run
              id: 9317
...
    RateLimited:
      description: Too many requests. Retry after the indicated number of seconds.
      headers:
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
            example: 30
        X-RateLimit-Limit:
          $ref: "#/components/headers/RateLimitLimit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/RateLimitRemaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/RateLimitReset"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: rate_limited
            message: Rate limit exceeded. Retry after 30 seconds.
            details:
              retry_after: 30
              limit: 120
              window: 60s
    Error:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # SCHEMAS
  schemas:
    Page:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          description: >
            Result items. The concrete type is defined by allOf composition
            in each endpoint response.
          items: {}
        pagination:
          type: object
          required: [limit, offset, total]
          properties:
            limit:
              type: integer
              minimum: 1
            offset:
              type: integer
              minimum: 0
            total:
              type: integer
              minimum: -1
              nullable: true
              description: >
                Total matching records. Null if count was not computed for this request.
                Pass ?count=true to force computation.
            next_offset:
              type: integer
              minimum: 0
              nullable: true
            truncated:
              type: boolean
              description: >
                Present and true when the result set was capped by an
                internal safety limit (e.g. status-filter on runs). When
                true, total may undercount the real number of matches.
...
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          maxLength: 100
          description: Machine-readable error code (snake_case)
          example: not_found
        message:
          type: string
          maxLength: 500
          description: Human-readable error summary
          example: Run 9317 not found.
        details:
          type: object
          additionalProperties: true
          description: >
            Structured context for the error. Always an object, never null.
            Empty object {} when no additional detail is available.
...
    QueueJob:
      type: object
      required: [run_id, status, platform, queued_at]
      properties:
        run_id:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [queued, running]
        platform:
          type: string
          enum: [linux, windows]
        queued_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
          nullable: true
        position:
          type: integer
          minimum: 1
          nullable: true
          description: Queue position. Null for jobs that are already running.
    Artifact:
      type: object
      required: [artifact_id, run_id, type, filename, content_type, storage_status]
      properties:
        artifact_id:
          type: string
          maxLength: 100
        run_id:
          type: integer
          minimum: 1
        sample_id:
          type: integer
          minimum: 1
          nullable: true
        type:
          type: string
          enum: [build_log, sample_output, expected_output, actual_output, diff, media_info, binary, coredump, combined_stdout]
        filename:
          type: string
          maxLength: 255
        content_type:
          type: string
          maxLength: 100
        size_bytes:
          type: integer
          minimum: 0
          nullable: true
        storage_status:
          type: string
          enum: [ok, degraded, missing]
          description: >
            ok = file verified in at least one storage backend.
            degraded = file exists but integrity or redundancy check failed.
            missing = file not found in any storage backend.
        download_url:
          type: string
          format: uri
          nullable: true
          description: >
            Only present and non-null when storage_status is ok or degraded.
            Always a verified URL. Null when storage_status=missing.